+++ src/connectClient/initiator/MewConnectInitiator.js
@@ -376,11 +376,11 @@ Keys
         this.requestIds.splice(idx, 1);
         debug('MESSAGE IDS KNOWN', this.requestIds)
       } else {
-        debug('**NO MESSAGE ID RECEIVED**');
+        debug('**NO MESSAGE ID RECEIVED : field present**');
         this.uiCommunicator(data.type, data.data);
       }
     } else {
-      debug('**NO MESSAGE ID RECEIVED**');
+      debug('**NO MESSAGE ID RECEIVED : field absent**');
       this.uiCommunicator(data.type, data.data);
     }
   }