+++ src/connectProvider/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import PopUpHandler from '../connectWindow/popUpHandler';
 import Initiator from '../connectClient/MewConnectInitiator';
 import Web3 from 'web3';
@@ -9,18 +10,19 @@ import EventEmitter from 'events';
 import EventNames from './web3Provider/web3-provider/events';
 import { Transaction } from 'ethereumjs-tx';
 import messageConstants from '../messageConstants';
-
 // import parseTokensData from './web3Provider/helpers/parseTokensData';
 import debugLogger from 'debug';
 import PopUpCreator from '../connectWindow/popUpCreator';
 
 const debugConnectionState = debugLogger('MEWconnect:connection-state');
+const debugErrors = debugLogger('MEWconnectError');
 
 let state = {
   wallet: null
 };
 const eventHub = new EventEmitter();
 let popUpCreator = {};
+const recentDataRecord = [];
 export default class Integration extends EventEmitter {
   constructor(options = {}) {
     super();
@@ -39,7 +41,14 @@ export default class Integration extends EventEmitter {
     this.initiator = new Initiator();
     this.popUpHandler = new PopUpHandler();
     this.connectionState = false;
-    this.chainIdMapping = Object.keys(Networks).reduce(
+    this.chainIdMapping = this.createChainMapping();
+    this.recent;
+    this.returnPromise = null;
+    popUpCreator = new PopUpCreator();
+  }
+
+  createChainMapping() {
+    return Object.keys(Networks).reduce(
       (acc, curr) => {
         if (Networks[curr].length === 0) return acc;
         acc.push({
@@ -54,8 +63,6 @@ export default class Integration extends EventEmitter {
       },
       [{ name: 'mainnet', chainId: 1, key: 'ETH' }]
     );
-    this.returnPromise = null;
-    popUpCreator = new PopUpCreator();
   }
 
   showNotifierDemo(details) {
@@ -194,10 +201,12 @@ export default class Integration extends EventEmitter {
           );
         }
       }
-
-      const parsedUrl = `${hostUrl.protocol}//${hostUrl.host}${
-        defaultNetwork.port ? ':' + defaultNetwork.port : ''
-      }${hostUrl.pathname}`;
+      console.log(hostUrl); // todo remove dev item
+      const parsedUrl = `${hostUrl.protocol}//${
+        hostUrl.hostname ? hostUrl.hostname : hostUrl.host
+      }${hostUrl.port ? ':' + hostUrl.port : ''}${
+        hostUrl.pathname ? hostUrl.pathname : ''
+      }`;
       state.enable = this.enable.bind(this);
       const web3Provider = new MEWProvider(
         parsedUrl,
@@ -220,6 +229,7 @@ export default class Integration extends EventEmitter {
       web3Provider.name = 'MewConnect';
       return web3Provider;
     } catch (e) {
+      debugErrors('makeWeb3Provider ERROR');
       if (chainError) {
         throw e;
       } else {
@@ -272,6 +282,7 @@ export default class Integration extends EventEmitter {
       console.warn('No connected wallet found');
       return true;
     } catch (e) {
+      debugErrors('disconnect ERROR');
       // eslint-disable-next-line
       console.error(e);
       return false;
@@ -281,6 +292,24 @@ export default class Integration extends EventEmitter {
   sign(tx) {
     if (state.wallet) {
       return state.wallet.signTransaction(tx);
+    }
+  }
+
+  checkDoubleSentTx(parsed) {
+    console.log(parsed); // todo remove dev item
+    const now = Date.now();
+    const entry = recentDataRecord.find(
+      entry => entry.tx.hash === parsed.tx.hash
+    );
+    console.log('-s-s-s-s-s-s-s-s-s-'); // todo remove dev item
+    console.log(entry); // todo remove dev item
+    console.log(now); // todo remove dev item
+    if (entry) {
+      console.log('-s-s-s-s-s-s-s-s-s-'); // todo remove dev item
+      console.log(entry); // todo remove dev item
+      console.log(now); // todo remove dev item
+    } else {
+      recentDataRecord.push({ parsed: parsed, timestamp: Date.now() });
     }
   }
 
@@ -298,9 +327,15 @@ export default class Integration extends EventEmitter {
           .signTransaction(tx)
           .then(_response => {
             this.popUpHandler.showNoticePersistentExit();
+            // if(this.checkDoubleSentTx(_response)){
+            //
+            // }
+            console.log(_response); // todo remove dev item
             resolve(_response);
+
           })
           .catch(err => {
+            debugErrors('sign transaction ERROR');
             this.popUpHandler.showNoticePersistentExit();
             state.wallet.errorHandler(err);
           });
@@ -323,6 +358,7 @@ export default class Integration extends EventEmitter {
             resolve(result);
           })
           .catch(() => {
+            debugErrors('sign message ERROR');
             this.popUpHandler.showNoticePersistentExit();
           });
       }
@@ -364,7 +400,8 @@ export default class Integration extends EventEmitter {
       this.lastHash = null;
       this.popUpHandler.showNotice(messageConstants.complete);
     });
-    eventHub.on('Error', () => {
+    eventHub.on('Error', e => {
+      debugErrors('SendTx:Error ERROR');
       if (this.lastHash !== null) {
         this.popUpHandler.showNotice(
           {