+++ src/connectProvider/web3Provider/WalletInterface.js
@@ -101,23 +101,6 @@ class WalletInterface {
     if (this.isPubOnly && typeof signer !== 'function')
       throw new Error('public key only wallets needs a signer');
     return new Promise((resolve, reject) => {
-      // if (!this.isPubOnly) {
-      //   const tx = new Transaction(txParams, {
-      //     common: commonGenerator(Networks.ETH)
-      //   });
-      //   const networkId = tx.getChainId();
-      //   tx.sign(this.privateKey);
-      //   const signedChainId = calculateChainIdFromV(tx.v);
-      //   if (signedChainId !== networkId)
-      //     throw new Error(
-      //       'Invalid networkId signature returned. Expected: ' +
-      //         networkId +
-      //         ', Got: ' +
-      //         signedChainId,
-      //       'InvalidNetworkId'
-      //     );
-      //   resolve(getSignTransactionObject(tx));
-      // } else {
         signer(txParams)
           .then(resolve)
           .catch(reject);