+++ src/connectProvider/web3Provider/web3-provider/methods/eth_sendTransaction.js
@@ -4,6 +4,9 @@ import { toPayload } from '../jsonrpc';
 import { getSanitizedTx } from './utils';
 import BigNumber from 'bignumber.js';
 import Misc from '../../helpers/misc';
+import debugLogger from 'debug';
+const debug = debugLogger('MEWconnectWeb3');
+const debugErrors = debugLogger('MEWconnectError');
 
 const setEvents = (promiObj, tx, eventHub) => {
   promiObj
@@ -44,7 +47,7 @@ export default async (
   getSanitizedTx(tx)
     .then(_tx => {
       eventHub.emit(EventNames.SHOW_TX_CONFIRM_MODAL, _tx, _response => {
-
+        debug('broadcasting', payload.method, _response.rawTransaction)
         const _promiObj = store.state.web3.eth.sendSignedTransaction(
           _response.rawTransaction
         );
@@ -63,6 +66,7 @@ export default async (
             res(null, toPayload(payload.id, hash));
           })
           .on('error', err => {
+            debugErrors('Error: eth_sendTransaction', err);
             res(err);
           });
         setEvents(_promiObj, _tx, eventHub);